\name{bambedObjDemo}
\alias{bambedObjDemo}
\docType{data}
\title{
    Demo data pre-stored for bambedObj.
}
\description{
    Pre-stored bambedObj data for demonstration purposes.
}
\usage{data(bambedObjDemo)}
\value{
    bambedObj demo data (list) pre-computed.
}
\details{
    Pre-computed using whole exome sequencing data of 46 HapMap samples.
}
\author{
    Yuchao Jiang \email{yuchaoj@wharton.upenn.edu}
}
\examples{
bamdir <- bambedObjDemo$bamdir
sampname <- bambedObjDemo$sampname
ref <- bambedObjDemo$ref
projectname <- bambedObjDemo$projectname
chr <- bambedObjDemo$chr
}
\keyword{datasets}