\name{coverageObjDemo}
\alias{coverageObjDemo}
\docType{data}
\title{
Demo data pre-stored for coverageObj.
}
\description{
Pre-stored coverageObj data for demonstration purposes.
}
\usage{data(coverageObjDemo)}
\value{
coverageObj demo data (list) pre-computed.
}
\details{
Pre-computed using whole exome sequencing data of 46 HapMap samples.
}
\author{
Yuchao Jiang \email{yuchaoj@wharton.upenn.edu}
}
\examples{
Y <- coverageObjDemo$Y
readlength <- coverageObjDemo$readlength
}
\keyword{datasets}