\name{gc_qc_codex2}
\alias{gc_qc_codex2}
\docType{data}
\title{
GC content for toy dataset for CODEX2.
}
\description{
Pre-stored GC content for demonstration purposes.
}
\usage{data(gc_qc_codex2)}
\value{
GC content demo data (vector) pre-computed.
}
\details{
Pre-computed from the toy dataset using CODEX.
}
\author{
Yuchao Jiang \email{yuchaoj@email.unc.edu}
}
\examples{
head(gc_qc_codex2)
}
\keyword{datasets}