\name{gcDemo}
\alias{gcDemo}
\docType{data}
\title{
Demo data pre-stored for GC content.
}
\description{
Pre-stored GC content data for demonstration purposes.
}
\usage{data(gcDemo)}
\value{
gc demo data (vector) pre-computed.
}
\details{
Pre-computed using whole exome sequencing data of 46 HapMap samples.
}
\author{
Yuchao Jiang \email{yuchaoj@wharton.upenn.edu}
}
\examples{
head(round(gcDemo, 2))
}
\keyword{datasets}