\name{mapp_hg19}
\alias{mapp_hg19}
\docType{data}
\title{
GRanges with pre-computed mappability for hg19.
}
\description{
GRanges object specifying exonic positions with mappabilities across the genome.
}
\usage{data(mapp_hg19)}
\value{
GRanges object with mappability for hg19.
}
\details{
GRanges object with mappability for hg19. Obtained from the ENCODE project 100mer.
}
\author{
Yuchao Jiang \email{yuchaoj@email.unc.edu}
}
\examples{
mapp_hg19
}
\keyword{datasets}