\name{mapp_hg38}
\alias{mapp_hg38}
\docType{data}
\title{
GRanges with pre-computed mappability for hg38.
}
\description{
GRanges object specifying exonic positions with mappabilities across the genome.
}
\usage{data(mapp_hg38)}
\value{
GRanges object with mappability for hg38.
}
\details{
GRanges object with mappability for hg38. Lifted over from the ENCODE project 100mer.
}
\author{
Yuchao Jiang \email{yuchaoj@email.unc.edu}
}
\examples{
mapp_hg38
}
\keyword{datasets}