\name{mappDemo}
\alias{mappDemo}
\docType{data}
\title{
Demo data pre-stored for mappability.
}
\description{
Pre-stored mappability data for demonstration purposes.
}
\usage{data(mappDemo)}
\value{
mapp demo data (vector) pre-computed.
}
\details{
Pre-computed using whole exome sequencing data of 46 HapMap samples.
}
\author{
Yuchao Jiang \email{yuchaoj@wharton.upenn.edu}
}
\examples{
head(round(mappDemo, 2))
}
\keyword{datasets}