\name{norm_index_codex2}
\alias{norm_index_codex2}
\docType{data}
\title{
Index for normal samples for toy dataset for CODEX2.
}
\description{
Normal index for demonstration purposes.
}
\usage{data(norm_index_codex2)}
\value{
Normal index demo data (vector).
}
\details{
Normal index for CODEX2's normalization procedure in the case-control setting.
}
\author{
Yuchao Jiang \email{yuchaoj@email.unc.edu}
}
\examples{
head(norm_index_codex2)
}
\keyword{datasets}