\name{ref_qc_demo}
\alias{ref_qc_demo}
\docType{data}
\title{
GC content for toy dataset for CODEX2.
}
\description{
Pre-stored GRanges object for demonstration.
}
\usage{data(ref_qc_demo)}
\value{
GRanges object pre-computed.
}
\details{
Pre-computed from the toy dataset. GC content is from real dataset from the 1000 Genomes Project WES data; start and end positions are manually constructed for demonstration purpose only.
}
\author{
Yuchao Jiang \email{yuchaoj@email.unc.edu}
}
\examples{
head(ref_qc_demo)
}
\keyword{datasets}