\name{Y_qc_codex2}
\alias{Y_qc_codex2}
\docType{data}
\title{
Read depth matrix for toy dataset for CODEX2.
}
\description{
Pre-stored read depth matrix (computed by CODEX with quality control) for demonstration purposes.
}
\usage{data(Y_qc_codex2)}
\value{
Read depth matrix pre-computed by CODEX.
}
\details{
Pre-computed from the toy dataset using CODEX.
}
\author{
Yuchao Jiang \email{yuchaoj@email.unc.edu}
}
\examples{
dim(Y_qc_codex2); length(gc_qc_codex2)
head(Y_qc_codex2)
}
\keyword{datasets}