\name{mapp_hg19}
\alias{mapp_hg19}
\docType{data}
\title{
GRanges with pre-computed mappability for hg19.
}
\description{
GRanges object specifying exonic positions whose 
mappabilities are pre-computed across the genome.
}
\usage{data(mapp_hg19)}
\value{
GRanges object with mappability pre-computed for hg19.
}
\details{
Genomic positions for pre-computed mappabilities. Method used is detailed in
\code{\link{getmapp}}.
}
\author{
Yuchao Jiang \email{yuchaoj@email.unc.edu}
}
\examples{
mapp_hg19
}
\keyword{datasets}